WSQ11
Called "Yo soy 196", a library is imported to C++ , allowing really big integer numbers to be converted to strings, then be compared to themselves backwards and through an evaluation of additions to see if they are part of the Lychrel numbers group of numbers, non-Lychrel palindromes, or just natural palindromes! I could not have done this without my friend juanpsantana from his GitHub and WithKnown pages (under the same name)! The library was made by @ken_bauer!