Quiz11 Question 2
By opening a file (user generated), the code should read the contents of the file and count how many times the word "banana" is written in the file. Due to problems when converting to a full lowercase with the locale library, this only works for all lower case or all upper case "banana" (or "BANANA").